- diff utils support added;

+++ app/src/main/java/com/cardee/inbox/chat/list/adapter/ChatDiffCallback.java
@@ -28,7 +28,7 @@ public class ChatDiffCallback extends DiffUtil.Callback {
 
     @Override
     public boolean areItemsTheSame(int oldItemPosition, int newItemPosition) {
-        return oldChatList.get(oldItemPosition).getChatId() == newChatList.get(newItemPosition).getChatId();
+        return oldChatList.get(oldItemPosition).getChatId().intValue() == newChatList.get(newItemPosition).getChatId().intValue();
     }
 
     @Override
